fix(opencode-plugin): keep helper exports off the published entrypoint (#6003)
