+++ integrations/mem0-plugin/.opencode-plugin/project.ts
@@ -1,10 +1,12 @@
+import {basename} from "path";
+
 /**
  * Project identity resolution for the Mem0 OpenCode plugin.
  *
  * The project id (`app_id`) scopes memories to a repo. We derive it from the
- * git remote so it is stable across clones, worktrees, and sub-directories —
- * falling back (in opencode-mem0.ts) to the git repo root dir name, then the
- * cwd. Keeping the parser pure makes the tricky remote formats testable.
+ * git remote so it is stable across clones, worktrees, and sub-directories,
+ * falling back to the git repo root dir name, then the cwd. Keeping the parser
+ * pure makes the tricky remote formats testable.
  */
 
 export type ProjectContext = {
@@ -20,6 +22,44 @@ export function selectActiveProjectPath(input: ProjectContext = {}): string {
   if (directory) return directory;
 
   return process.cwd();
+}
+
+type ShellCommand = {
+  cwd?: (path: string) => ShellCommand;
+  quiet: () => Promise<{ stdout: { toString(): string } }>;
+};
+
+function commandInProject(command: ShellCommand, projectPath: string): ShellCommand {
+  if (typeof command.cwd === "function") return command.cwd(projectPath);
+  return command;
+}
+
+export async function getProjectId($: any, projectPath: string): Promise<string> {
+  if (process.env.MEM0_APP_ID) return process.env.MEM0_APP_ID;
+  try {
+    const r = await commandInProject($`git remote get-url origin`, projectPath).quiet();
+    const project = parseProjectFromRemote(r.stdout.toString());
+    if (project) return project;
+  } catch {
+  }
+  try {
+    const r = await commandInProject($`git rev-parse --show-toplevel`, projectPath).quiet();
+    const top = r.stdout.toString().trim();
+    if (top) return basename(top);
+  } catch {
+  }
+  const selectedBasename = basename(projectPath);
+  if (selectedBasename) return selectedBasename;
+  return basename(process.cwd());
+}
+
+export async function getBranch($: any, projectPath: string): Promise<string> {
+  try {
+    const r = await commandInProject($`git branch --show-current`, projectPath).quiet();
+    return r.stdout.toString().trim() || "main";
+  } catch {
+  }
+  return "main";
 }
 
 /**
